+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes will be documented in this file.
 
 ### Fixed
 
+* WSAPI modify request appeared successful but was rejected (#586)
 * New URLs for market data streams (#584)
 * New URLs for user-stream (#583)